pyproject.toml
[project]
name = "tgubers"
version = "0.1.0"
description = "Parse and analyze content from Telegram groups of celebrities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "telethon>=1.36",
    "python-dotenv>=1.0",
    "ipython>=9.10.0",
    "tqdm>=4.67.3",
]
